+++ types/src/temperatureUtils.d.ts
@@ -1,0 +1,7 @@
+export declare enum TemperatureUnit {
+    Celsius = "Celsius",
+    Fahrenheit = "Fahrenheit",
+    Kelvin = "Kelvin"
+}
+export declare const zeroCelsiusInKelvin = 273.15;
+export declare function convertTemperature(value: number, fromUnit: TemperatureUnit, toUnit: TemperatureUnit): number;